moved combine_grids function from graphics to utils
